Address PR comments. Cleanup
